회원탈퇴 시 비밀번호 확인 기능 구현 및 article index 하트토글키 구현

+++ sub_pair2/accounts/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render, redirect
-from .forms import CustomUserCreationForm, CustomUserChangeForm
+from .forms import CustomUserCreationForm, CustomUserChangeForm, CheckPasswordForm
 from django.contrib.auth.forms import AuthenticationForm, PasswordChangeForm
 from django.contrib.auth import login as auth_login
 from django.contrib.auth import logout as auth_logout
 from django.contrib.auth import get_user_model, update_session_auth_hash
+from django.contrib import messages
 
 
 # Create your views here.
@@ -16,7 +17,7 @@ def signup(request):
         if form.is_valid():
             user = form.save()
             auth_login(request, user)
-            return redirect('accounts:index')
+            return redirect('articles:index')
     else:
         form = CustomUserCreationForm()
     
@@ -74,10 +75,19 @@ def change_password(request):
 
     return render(request, 'accounts/change_password.html', context)
 
-def delete(request):
-    request.user.delete()
-    auth_logout(request)
-    return redirect('accounts:index')
+def delete_confirm(request):
+    if request.method == 'POST':
+        password_form = CheckPasswordForm(request.user, request.POST)
+        
+        if password_form.is_valid():
+            request.user.delete()
+            logout(request)
+            messages.success(request, "회원탈퇴가 완료되었습니다.")
+            return redirect('accounts:index')
+    else:
+        password_form = CheckPasswordForm(request.user)
+
+    return render(request, 'accounts/delete_confirm.html', {'password_form':password_form})
 
 def detail(request, pk):
     user = get_user_model().objects.get(pk=pk)
@@ -85,3 +95,4 @@ def detail(request, pk):
         'user': user
     }
     return render(request, 'accounts/detail.html', context)
+
